Improvements to option and filename handling
Leanpub export options are now handled through the existing
:options-alist mechanism in the exporter definition, instead of
manually extracting the options from the file. The options have been
renamed:

 LEANPUB_EXPORT_OUTPUT_DIR to LEANPUB_OUTPUT_DIR
 LEANPUB_EXPORT_WRITE_SUBSET to LEANPUB_WRITE_SUBSET

Also, the generated output filenames now use the NAME, ID or CUSTOM_ID
attribute of the corresponding headline, if they exist, otherwise the
name is generated by normalizing the title. The EXPORT_FILE_NAME
property of each exported headline is updated if it does not exist, or
if it does not match the generated filename.